requirements.txt
Babel==2.3.4
blinker==1.3
cairocffi==0.7.2
CairoSVG==1.0.22
cffi==1.8.2
cssmin==0.2.0
cssselect==0.9.2
demjson==2.2.4
Flask==0.10.1
Flask-Assets==0.10
Flask-Babel==0.9
Flask-WeasyPrint==0.5
Flask-WTF==0.9.3
Flask-Limiter==0.9.3
freezegun==0.3.7
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.7.3
libsass==0.8.2
lxml==3.6.4
MarkupSafe==0.23
misaka==2.0.0
pbr==1.10.0
pg8000==1.10.3
pycparser==2.14
pycrypto==2.6.1
Pyphen==0.9.4
python-dateutil==2.4.0
pytz==2017.2
requests==2.5.1
Werkzeug==0.9.6
WTForms==1.0.5
